requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.6
django-countries==7.6.1
django-filter==24.2
django-postmark==0.1.6
djangorestframework==3.15.1
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
packaging==24.0
pillow==10.3.0
postmarker==1.0
psycopg2-binary==2.9.9
pyparsing==3.1.2
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.1
requests==2.32.3
sendgrid==6.11.0
six==1.16.0
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
urllib3==2.2.1
whitenoise==6.6.0
cloudinary 
django-cloudinary-storage
stripe
# cloudinary==1.40.0
# django-cloudinary-storage==0.3.0
# stripe==10.0.0
